requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.7
djangorestframework==3.15.2
httpie==3.2.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
psycopg==3.2.1
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==71.0.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
